Keep app_protocol as-is
