+++ src/pkg/cli/compose/fixup.go
@@ -11,6 +11,7 @@ import (
 	"github.com/DefangLabs/defang/src/pkg/term"
 	defangv1 "github.com/DefangLabs/defang/src/protos/io/defang/v1"
 	"github.com/compose-spec/compose-go/v2/types"
+	composeTypes "github.com/compose-spec/compose-go/v2/types"
 )
 
 func FixupServices(ctx context.Context, provider client.Provider, project *types.Project, upload UploadMode) error {
@@ -25,8 +26,7 @@ func FixupServices(ctx context.Context, provider client.Provider, project *types
 	for _, svccfg := range project.Services {
 		// Fixup ports (which affects service name replacement by ReplaceServiceNameWithDNS)
 		for i, port := range svccfg.Ports {
-			fixupPort(&port)
-			svccfg.Ports[i] = port
+			svccfg.Ports[i] = fixupPort(port)
 		}
 	}
 
@@ -305,4 +305,31 @@ func getImageRepo(imageRepo string) string {
 	image := strings.ToLower(imageRepo)
 	image, _, _ = strings.Cut(image, ":")
 	return image
+}
+
+func fixupPort(port composeTypes.ServicePortConfig) composeTypes.ServicePortConfig {
+	switch port.Mode {
+	case "":
+		term.Warnf("port %d: no 'mode' was specified; defaulting to 'ingress' (add 'mode: ingress' to silence)", port.Target)
+		fallthrough
+	case Mode_INGRESS:
+		// This code is unnecessarily complex because compose-go silently converts short `ports:` syntax to ingress+tcp
+		if port.Protocol != Protocol_UDP {
+			if port.Published != "" {
+				term.Debugf("port %d: ignoring 'published: %s' in 'ingress' mode", port.Target, port.Published)
+			}
+			if port.AppProtocol == "" {
+				// TCP ingress is not supported; assuming HTTP (add 'app_protocol: http' to silence)"
+				port.AppProtocol = "http"
+			}
+		} else {
+			term.Warnf("port %d: UDP ports default to 'host' mode (add 'mode: host' to silence)", port.Target)
+			port.Mode = Mode_HOST
+		}
+	case Mode_HOST:
+		// no-op
+	default:
+		panic(fmt.Sprintf("port %d: 'mode' should have been validated to be one of [host ingress] but got: %v", port.Target, port.Mode))
+	}
+	return port
 }
